Function sumi::common::write_common_rna_file

pub fn write_common_rna_file(
    rna_names: Vec<String>,
    rna_info: Vec<HashMap<String, i32>>,
    rpm_info: Vec<HashMap<String, f64>>,
    sample_names: Vec<String>,
    reference_name: &str
)

Function to write RNAs common to all samples to two output files: one for counts and one for RPM values. Each row in the output files represents a RNA and each subsequent column represents the counts (or RPM values) for that RNA in one sample.

Arguments

rna_names - A vector of RNA names that are common to all samples.

rna_info - A vector of HashMaps where each HashMap contains the RNA counts for one sample.

rpm_info - A vector of HashMaps where each HashMap contains the RPM values for the RNAs in one sample.

sample_names - A vector of sample names.

reference_name - Name of the reference.

Returns

Nothing, the function writes to output files.

Example

let rna_names = vec!["mirna1", "mirna2", "mirna3"];
let rna_info = vec![
    hashmap!{ "mirna1" => 10, "mirna2" => 20, "mirna3" => 30 },
    hashmap!{ "mirna1" => 40, "mirna2" => 50, "mirna3" => 60 },
];
let rpm_info = vec![
    hashmap!{ "mirna1" => 0.1, "mirna2" => 0.2, "mirna3" => 0.3 },
    hashmap!{ "mirna1" => 0.4, "mirna2" => 0.5, "mirna3" => 0.6 },
];
let sample_names = vec!["sample1", "sample2"];

write_common_rna_file(
    rna_names,
    rna_info,
    rpm_info,
    sample_names,
    "miRNA",
);