Function sumi::extract_umis::regex_extraction

pub fn regex_extraction(
    read_sequence: String,
    record: Record,
    umi_delineator: &str,
    writer: &mut Writer<File>,
    umi_regex: &Regex,
    is_qiagen: bool,
    is_3p: bool,
    start: usize,
    adapter_mismatch: bool
)

Extract Unique Molecular Identifiers (UMIs) from a given sequence using regex.

The UMIs are searched using a regex pattern related to known adapter sequences and can be located on either the 5’ or 3’ end. The function is able to process sequences from both gzipped and uncompressed FASTQ files.

Arguments

read_sequence - The DNA sequence from the FASTQ record.

record - The FASTQ record from which the UMI and other details are extracted.

umi_delineator - Delimiter used between appended UMI and read name. The default is “_” as is used in UMI-tools.

writer - A mutable reference to the FASTQ writer for writing out the UMI-trimmed sequences.

umi_regex - The regex pattern used to identify and capture the UMI in the read_sequence.

is_qiagen - A flag indicating a specific trimming behavior.

is_3p - A flag indicating whether the UMI is on the 3’ end of the sequence.

start - The index where the regex first matches.

adapter_mismatch - A flag indicating whether to allow mismatches and indels during alignment.

Example

let mut writer = Writer::new(File::create(format!("{}.processed.fastq", "sample_1")).unwrap());
 
regex_extraction("GCATGCTACTGCTCGAT", record, "_", &mut writer, &"(.{12}$)", false, true, 35, false);
// This will write to `writer` the UMI-trimmed sequences with UMI appended to the record's description.

Note: The UMI is extracted using the provided umi_regex. This function is flexible to different UMI positions and patterns, unlike the previous version which expected UMIs to be the first 12 characters of the sequence.

It’s important to ensure that the regex pattern provided accurately captures the UMI without unintentionally capturing other portions of the sequence.