Function sumi::graph::build_substring_index

pub fn build_substring_index(
    umis: &[Vec<u8>],
    slice_size: usize
) -> HashMap<Vec<u8>, HashSet<usize>>

Construct a substring index from a slice of UMIs.

This function creates an index that maps each unique substring of a UMI to a set of UMIs that contain that substring. The size of the substring is determined by slice_size.

The index is used to quickly find potential neighbor UMIs that may be within a certain edit distance, without having to compute the distance between all pairs of UMIs.

Arguments

umis - A slice of UMIs, where each UMI is a byte array.

slice_size - The size of the substring to use for the index. Each UMI is divided into chunks of this size (from left to right, and without overlap between chunks), and each chunk is added to the index.

Returns

A hashmap where the keys are substrings of UMIs, and the values are sets of indices. Each index in a set corresponds to a UMI that contains the substring.

Example

let umis: Vec<Vec<u8>> = vec![b"ACGT".to_vec(), b"TGCA".to_vec(), b"GTAC".to_vec(), b"CATG".to_vec()];
let slice_size: usize = 2;
let index = build_substring_index(&umis, slice_size);

In the above example, the index would map “AC” to {0}, “CG” to {0}, “GT” to {0, 2}, “TG” to {1, 3}, “CA” to {1, 3}, “GA” to {2}, and “AT” to {3}.