Function sumi::graph::deduplicate_bam

pub fn deduplicate_bam(
    input_bam_file: &str,
    sample_name: &String,
    true_umis: HashSet<Vec<u8>>
) -> Result<(), Box<dyn Error>>

Deduplicate a BAM file using a set of unique molecular identifiers (UMIs) and their mapping positions.

This function reads a BAM file and creates a deduplicated version of it based on a provided set of unique molecular identifiers (UMIs) and their mapping positions. The UMI associated with each read in the BAM file is expected to be appended to the read’s name, prefixed with an underscore (‘_’).

A read is written to the output file if its UMI is in the set of provided UMIs and its combination of UMI and mapping position hasn’t been observed before. If a read’s UMI is not in the set or the UMI-mapping position combination has been observed before, the read is discarded.

Arguments

input_bam_file - A string representing the path to the input BAM file.

sample_name - A string representing the sample name, which is used to name the output BAM file.

true_umis - A HashSet of UMIs (Vec<u8>), which are considered to be the true UMIs.

Output

This function doesn’t return any value. Instead, it writes to an output BAM file named {sample_name}.dedup.bam, where {sample_name} is the provided sample name.

Example

// assume `true_umis` is predefined
let result = deduplicate_bam("input.bam", "sample1", true_umis);
// handle the error, if any
match result {
    Ok(_) => println!("Deduplication completed successfully"),
    Err(e) => eprintln!("An error occurred: {}", e),
}