Function sumi::graph::umi_graph

pub fn umi_graph(
    umi_count_dict: &HashMap<Vec<u8>, u32>,
    max_edits: u32,
    use_levenshtein: bool
) -> (Graph<Vec<u8>, u32>, HashMap<NodeIndex, (Vec<u8>, u32)>)

Create a graph from a dictionary of UMIs and their counts.

This function constructs a graph where each node represents a unique UMI and each edge represents a edit distance less than or equal to max_edits between two UMIs. The edge creation is based on the condition that the count of one UMI should be at least five times the count of the other UMI minus 1. If the condition is met, an edge is added between the two UMIs.

The function implements a slightly modified directional graph algorithm that allows for a edit Distance of 1 between UMIs, see Fu, Y., et al, (2018). Elimination of PCR duplicates in RNA-seq and small RNA-seq using unique molecular identifiers. https://doi.org/10.1186/s12864-018-4933-1. The original algorithm uses a two fold count threshold. This version of the algorithm uses a substring indexing strategy to efficiently find potential neighbors for each UMI before computing the edit distance, hence improving the performance for large datasets.

Note that Levenshtein distance could be used here, but it generally takes longer to run and has a negligible effect on UMI deduplication.

Arguments

umi_count_dict - A hashmap where the key is the UMI (as a byte array) and the value is its count.

max_edits - The maximum edit distance allowed to consider two UMIs as neighbors (i.e., to add an edge between them).

use_levenshtein - Whether to use the Levenshtein distance or the default Hamming distance as the edit distance metric.

Returns

A tuple consisting of the UMI graph and a hashmap containing the node attributes. The graph is of type Graph<Vec<u8>, u32>, and the hashmap key is a NodeIndex which points to a tuple of UMI and its count.

Example

let umi_count_dict: HashMap<Vec<u8>, u32> = HashMap::new();
// populate umi_count_dict
let max_edits: u32 = 1;
let (graph, node_attributes) = umi_graph(&umi_count_dict, max_edits);