Function sumi::quality::get_q_score_probability

pub fn get_q_score_probability(q_score: &[u8]) -> f32

Get the total error probability from a record’s quality scores

This function accepts ASCII-encoded quality score and produces a scalar value that estimates the ‘total error probability’ for that record. It means that if you want to calculate average quality score, you don’t just sum up all the phred scores and find the average, but rather consider the probability distribution of the scores.

The Phred score for a base Q is calculated as Q = ASCII value of quality score - 33. The error probability P for that base is then calculated as P = 10^(-Q/10). Then these probabilities are summed up for all the bases to derive total error.

Arguments

q_score - The quality scores of the current record as an ASCII-encoded byte string.

Returns

The total error probability for the record.

References

Illumina’s explanation on quality scores and their corresponding error rates: https://www.illumina.com/science/technology/next-generation-sequencing/plan-experiments/quality-scores.html