Function sumi::rna_counts::generate_rna_counts

pub fn generate_rna_counts(
    input_sam_file: &str,
    sample_name: &String,
    reference_name: &str
) -> Result<HashMap<String, String>, Box<dyn Error>>

This function counts the occurrence of each RNA in an input SAM file, computes their reads per million (RPM),and writes this information to an output CSV file. The CSV file will contain the name of each unique RNA, its count,and RPM for the given sample.

The output CSV file will have the format "{sample_name}_{reference_name}_counts.csv" and will be located in the directory where the program is run.

Returns

A Result which is Ok if the function executed successfully. It is Err if an error occurred, with a description of the error.

Arguments

input_sam_file: The path to the input SAM file.

sample_name: The name of the sample.

reference_name: The name of the bowtie2 reference.

Example

let sam_file = "sample1.sam";
let reference_name = "miRNA"

let _ = generate_rna_counts(sam_file, "sample1", reference_name);
// This will generate a file "sample1_miRNA_counts.csv" with each miRNA found and
// their counts and CPM.