Function sumi::rna_counts::get_percent_alignment

pub fn get_percent_alignment(
    sample_name: &String,
    num_threads: u8,
    reference_name: &str,
    count: u64
) -> Result<f64, Box<dyn Error>>

Get the percentage of reads that aligned to the specified reference RNA type.

Divide reads aligned with total reads to calculate percent alignment.

/// # Arguments

count - The total number of reads analyzed.

sample_name - The name of the sample.

reference_name - Path to the bowtie2 reference.

num_threads - Number of threads to use for alignment and conversion processes.

Returns

The percent each fastq aligned to the specified RNA type.

Example

let reference_name = "miRNA"
let count = 12345678
 
let percent_alignment = get_percent_alignment("sample1", 4, reference_name, count);
println!("The fastq contains {} percent {}.", percent_alignment, reference_name);