Function sumi::rna_counts::is_sam_file_empty

pub fn is_sam_file_empty(sam_file: &str) -> Result<(), Error>

Check a SAM file to determine if it contains alignment data.

The function reads through the SAM file line by line, looking for any line that does not start with ‘@’, which would indicate it contains alignment data. If such a line is found, the function will return Ok(()), signaling the file does contain alignments.

If the end of the file is reached without encountering any alignment data (i.e., all lines start with ‘@’), an error is returned with the message “SAM file is empty”.

Arguments

sam_file - A string slice that holds the path to the SAM file.

Returns

Ok(()) if the SAM file contains alignment data.

Err(io::Error) if the SAM file does not contain alignment data or if there was an error reading the file.

Errors

This function will return an error if there was an issue reading the file or if the file does not contain alignment data.

Example

let result = is_sam_file_empty("/path/to/your/file.sam");
match result {
    Ok(_) => println!("The SAM file contains alignments."),
    Err(e) => eprintln!("An error occurred: {}", e),
}