Function sumi::sample::count_reads

pub fn count_reads(input_fastq: &str) -> Result<u64, Box<dyn Error>>

Count the number of reads in a FASTQ file.

Arguments

input_fastq - A string slice that holds the name of the FASTQ file to read.

Returns

This function will return the number of reads in the input FASTQ file as a Result<u64, Box<dyn std::error::Error>>. If an error occurs while reading the file, the error will be returned instead.

Examples

let num_reads = count_reads("my_file.fastq")?;
println!("The file contains {} reads.", num_reads);